Rename Bob.Release to Bob.ReleaseTasks and drop backfill task
Match the hexpm ReleaseTasks naming convention. Remove the backfill/0
release task: it hand-started a curated subset of apps and authenticated
Docker Hub itself, but never started the RateLimiter supervision child the
pager needs, so reconcile would crash on the first page. Run the backfill
on the already-running node instead, which reuses the full supervision
tree:

    bin/bob rpc "Task.Supervisor.start_child(Bob.Tasks, fn -> Bob.Reconcile.backfill() end)"
